x86: Add early SHA support for Secure Launch early measurements
The SHA algorithms are necessary to measure configuration information into
the TPM as early as possible before using the values. This implementation
uses the established approach of #including the SHA libraries directly in
the code since the compressed kernel is not uncompressed at this point.

The SHA code here has its origins in the code from the main kernel. That
code could not be pulled directly into the setup portion of the compressed
kernel because of other dependencies it pulls in. The result is this is a
modified copy of that code that still leverages the core SHA algorithms.

Signed-off-by: Daniel P. Smith <dpsmith@apertussolutions.com>
Signed-off-by: Ross Philipson <ross.philipson@oracle.com>

Version 2
==========================================================

 - With newer GCC packages, the exports in the sha256.c code cause problem in
   the PIE compressed kernel. This conditionally removes the exporting.

Conflicts:
	arch/x86/boot/compressed/Makefile