[patch] x86: Add early SHA1 support for Secure Launch early measurements
The SHA1 algorithm is necessary to measure configuration information into
the TPM as early as possible before using the values. This implementation
uses the established approach of #including the SHA1 library directly in
the code since the compressed kernel is not uncompressed at this point.

The code here has its origins in the code in include/crypto/sha1_base.h.
That code could not be pulled directly into the setup portion of the
compressed kernel because of other dependencies it pulls in. So this is
a modified copy of that code that still leverages the core SHA1
algorithm.

Signed-off-by: Ross Philipson <ross.philipson@oracle.com>
Reviewed-by: Alexandre Chartre <alexandre.chartre@oracle.com>