x86: Secure Launch kernel early boot stub
The Secure Launch (SL) stub provides the entry point for Intel TXT (and
later AMD SKINIT) to vector to during the late launch. The symbol
sl_stub_entry is that entry point and its offset into the kernel is
conveyed to the launching code using the MLE (Measured Launch
Environment) header in the structure named mle_header. The offset of the
MLE header is set in the kernel_info. The routine sl_stub contains the
very early late launch setup code responsible for setting up the basic
environment to allow the normal kernel startup_32 code to proceed. It is
also responsible for properly waking and handling the APs on Intel
platforms. The routine sl_main which runs after entering 64b mode is
responsible for measuring configuration and module information before
it is used like the boot params, the kernel command line, the TXT heap,
an external initramfs, etc.

Signed-off-by: Ross Philipson <ross.philipson@oracle.com>

Version 2
==========================================================

 - Allow use of alternate PCR 19 and 20 for post ACM measurements.
 - Use of PCRs 19 and 20 can be set independently via Kconfig.

 - Validate OS-MLE buffers before using them.
   This protects agains malicious pre-launch software placing the
   buffers over the MLE image and using this to gain control.

 - Check that the OS-SINIT version is 6 or greater
   to allow for safely locating and using he TPM 2.1 logging
   struct.

 - Add integer overflow/carry checks to buffer base and size values.
   The values passed to the MLE from the pre-launch environment
   are a base and a size. When added together an attacker could
   cause an integer overflow which needs to be detected.

 - Validate the event log container size

 - Ensure that the event log container size is less than or equal to the
   overall event log allocatd buffer.

 - Validate kernel relocation and decompression regions.

 - The compress kernel relocation and the decompression regions have to be <
   4GB and protected by the low PMR.

Conflicts:
	arch/x86/boot/compressed/head_64.S