[revision] for {d9efa1d76ba1}
Add ifdef CONFIG_SECURE_LAUNCH to kernel_info

This was unfortunately discovered after the RFC went out and it
prevents building w/ CONFIG_SECURE_LAUNCH disabled.

Signed-off-by: Ross Philipson <ross.philipson@oracle.com>

[revision] for {447296a93c21c0e519d87fe5df16685c86d6334e}

Fix long standing crash seen on Oracle servers

The problem was the AP stacks were not being moved to the AP wake block.
The kernel was uncompressed in place so %esp for the APs pointed to memroy
within the uncompressed kernel.

Signed-off-by: Ross Philipson <ross.philipson@oracle.com>

[revision] for {447296a93c21c0e519d87fe5df16685c86d6334e}

Add AP wake block size and size check in the early stub

Signed-off-by: Ross Philipson <ross.philipson@oracle.com>

[revision] for {447296a93c21c0e519d87fe5df16685c86d6334e}

Some adjustments to the OS-MLE structure in the TXT heap

Signed-off-by: Ross Philipson <ross.philipson@oracle.com>

[revision] for {447296a93c21c0e519d87fe5df16685c86d6334e}

Allow the TPM event log to reside anywhere

Added a base and size field so the TPM event log can be allocated anywhere.
Note this does not preclude it from still residing in the TXT heap if
desired.

This commit will have some issues with the AMD patch when rebasing the
branch next.

Signed-off-by: Ross Philipson <ross.philipson@oracle.com>

[revision] for {447296a93c21c0e519d87fe5df16685c86d6334e}

Fix copyright statements.

Signed-off-by: Ross Philipson <ross.philipson@oracle.com>

[revision] for {447296a93c21c0e519d87fe5df16685c86d6334e}

Remove linker check for the AP wake block size

This is no longer needed since it is now checked at runtime.

Signed-off-by: Ross Philipson <ross.philipson@oracle.com>

[revision] for {447296a93c21c0e519d87fe5df16685c86d6334e}

Minor modifications to the main SL header from code review

Note this commit will have issues moving it back past the AMD
patches.

Signed-off-by: Ross Philipson <ross.philipson@oracle.com>

[revision] for {447296a93c21c0e519d87fe5df16685c86d6334e}

Minor modifications to the early and late stub code from review

Signed-off-by: Ross Philipson <ross.philipson@oracle.com>

[revision] for {447296a93c21c0e519d87fe5df16685c86d6334e}

Switch to using %ebx for the base pointer to the kernel image

The code was originally using %ebp but this is not ideal. The
default addressing for %ebp is through %ss and it is not clear
this segment is valid. Also some operations related to %ss do
not behave as they do in other data segments.

Signed-off-by: Ross Philipson <ross.philipson@oracle.com>

[revision] for {447296a93c21c0e519d87fe5df16685c86d6334e}

Minor fixes to early stub code. Note this will probably cause
issues during the rebase.

Signed-off-by: Ross Philipson <ross.philipson@oracle.com>