[revision] for {447296a93c21c0e519d87fe5df16685c86d6334e}
Allow the TPM event log to reside anywhere

Added a base and size field so the TPM event log can be allocated anywhere.
Note this does not preclude it from still residing in the TXT heap if
desired.

This commit will have some issues with the AMD patch when rebasing the
branch next.

Signed-off-by: Ross Philipson <ross.philipson@oracle.com>