[revision] for {447296a93c21c0e519d87fe5df16685c86d6334e}
Switch to using %ebx for the base pointer to the kernel image

The code was originally using %ebp but this is not ideal. The
default addressing for %ebp is through %ss and it is not clear
this segment is valid. Also some operations related to %ss do
not behave as they do in other data segments.

Signed-off-by: Ross Philipson <ross.philipson@oracle.com>