x86: Prepare CPUs for post SKINIT launch
The SKINIT instruction disables the GIF and it must be re-enabled
on the BSP and APs as they are started. Since enabling GIT also
re-enables NMIs, it should be done after a valid IDT is loaded for
each CPU.

The #INITs were already sent to the APs before SKINIT and it should not
be done again by the kernel.

Signed-off-by: Ross Philipson <ross.philipson@oracle.com>
Signed-off-by: Sergii Dmytruk <sergii.dmytruk@3mdeb.com>