x86: Extend PCR measurements in TPM during late initialization
The initial measurements post SKINIT launch are taken in the early
Secure Launch stub code. There is no TPM access available at this
point so the measurements are written to the TPM event log. The
Secure Launch late platform module has access to the mainline TPM
driver. The measurements are located in the event log, read and
written to the TPM PCRs.

Signed-off-by: Ross Philipson <ross.philipson@oracle.com>
Signed-off-by: Sergii Dmytruk <sergii.dmytruk@3mdeb.com>