x86: Place TXT MLE header in the kernel_info section
The MLE (measured launch environment) header must be locatable by the
boot loader and TXT must be setup to do a launch with this header's
location. While the offset to the kernel_info structure does not need
to be at a fixed offset, the offsets in the header must be relative
offsets from the start of the setup kernel. The support in the linker
file achieves this.

TODO when squashing:

 - Remove patch 1 as a commit.
 - Make this a new patch that comes before current patch 8.
 - Remove MLE header setup from patch 8.
 - Delete this comment.
 - Add a suggested by line for Ard.

Signed-off-by: Ross Philipson <ross.philipson@oracle.com>