[revision] for {adbd991cff7a2fc2803a4e1c9ef2100c2683055c}
This is an empty commit to indicate a change to a commit
message during the rebase. The commit should be dropped during
the rebase.

Add the following as a second paragraph:

Intel VT-d/IOMMU hardware provides special registers called Protected
Memory Regions (PMRs) that allow all memory to be protected from
DMA during a TXT DRTM launch. This coverage is validated during the
late setup process to ensure DMA protection is in place prior to
the IOMMUs being initialized and configured by the mainline kernel.
See the Intel Trusted Execution Technology - Measured Launch Environment
Developer’s Guide for more details.

Signed-off-by: Ross Philipson <ross.philipson@oracle.com>