Commit message changes for v13 submission
This is an empty commit listing all the updated commit messages
that will go in the patch set.

x86: Secure Launch Resource Table header file
----------------------------------------------------------------

The Secure Launch Specification is an architectural agnostic, software neutral
API/ABI maintainted by the TrenchBoot project. It's function is to allow any
compliant boot loader to communicate the pre-launch configuration to any
compliant post-launch kernel. This API/ABI is referred to as the Secure Launch
Resource Table by the specification.

x86: Secure Launch main header file
----------------------------------------------------------------

Introduce the main Secure Launch header file used in the early SL stub
and the early setup code.

This header file contains the following categories
 - Secure Launch implementation specific structures and definitions.
 - Intel TXT architecture specific DRTM structures, definitions and functions
   used by Secure Launch.
 - DRTM TPM event logging definitions and helper functions.

x86/msr: Add variable MTRR base/mask and x2apic ID registers
----------------------------------------------------------------

Add the MSR values required by Secure Launch to locate particular CPU cores
during application processor (AP) startup, and restore the MTRR state after
an Intel TXT launch.

x86/boot: Place TXT MLE header in the kernel_info section
----------------------------------------------------------------

The Measured Launch Environment (MLE) header must be locatable by the
boot loader and Intel TXT must be setup to do a launch with this header's
location. While the offset to the kernel_info structure does not need
to be at a fixed offset, the offsets in the header must be relative
offsets from the start of the setup kernel. Note that from the viewpoint
of the prelaunch phase and TXT, the setup kernel image as loaded into
memory is the MLE image.

The changes to the linker file achieve this by making available the
offset values which are updated in the MLE header structure. The following
are the needed offsets from the beginning of the setup kernel image:

 - kernel_info_offset: Offset of the main kernel_info structure.
 - mle_header_offset: Offset of the MLE header structure.
 - sl_stub_entry_offset: Offset of the Secure Launch initial entry point.
 - _edata_offset: Offset of the _edata label used as the end of the MLE image.

x86: Secure Launch kernel early boot stub
----------------------------------------------------------------

The Secure Launch (SL) stub provides the entry point for Intel TXT to
jump to during the dynamic launch. The symbol sl_stub_entry is that entry
point and its offset into the kernel is conveyed to the launching code using
the Measured Launch Environment (MLE) header in the structure named mle_header.
The offset of the MLE header is set in the kernel_info.

The routine sl_stub contains the very early dynamic launch setup code
responsible for setting up the basic operating environment to allow the normal
kernel startup_32 code to proceed. It is also responsible for properly waking
and handling the APs on Intel platforms.

The routine sl_main which runs after entering 64b mode in the setup kernel. It
is responsible for measuring configuration and module information before
it is used. An example of entities measured on Intel x86 are the boot params,
the kernel command line, the TXT heap, any external initramfs, etc. In addition
this routine does some early setup and validation of the environment like
locating the TPM event log and validating the location of various buffers to
ensure they are protected and not overlapping.

x86: Secure Launch kernel late boot stub
----------------------------------------------------------------

slaunch_setup -> slaunch_setup()

Last para:

In addition this routine reserves key memory regions used by Secure Launch
(e.g. the TXT heap, AP startup block etc) as well as fetching values needed later
from the TXT heap and SLRT.

Signed-off-by: Ross Philipson <ross.philipson@oracle.com>