tpm-buf: Remove chip parameter from tpm_buf_append_handle
Remove chip parameter from tpm_buf_append_handle() in order to maintain
decoupled state with tpm-buf. This is mandatory change in order to re-use
the module in early boot code of Trenchboot, and the binding itself brings
no benefit. Use WARN like in other functions, as the error condition can
happen only as a net effect of a trivial programming mistake.

Signed-off-by: Jarkko Sakkinen <jarkko.sakkinen@opinsys.com>
Message-ID: <20250929194832.2913286-9-jarkko@kernel.org>
Signed-off-by: Ross Philipson <ross.philipson@oracle.com>